Add support for whitelisting folders
Users have media files scattered throughout their device, most of which
they don't want cluttering up their music library. A solution to this
was proposed in issue 22, where one/multiple folders could be specified
(a whitelist) and only files in those folder trees would make up the
library.

This commit is a hacky proof of concept for a solution to this. It works
as you might expect with the following limitations:

* white list is hard-coded in (though there are setters/getters for it)
* The existing genre hack was further butchered to accomodate this hack
 and will need to be worked on further. There will be a limit on how
 clean it can be, but it isn't anywhere near there yet.

This is a work in progress. It has only been desk-tested. It also needs to
be cleaned up, and at point 1) addressed, probably through a preference that
takes a list of folders.